[FEAT] 대기열 토큰 관리에 Redis 적용

+++ src/main/java/kr/hhplus/be/server/module/auth/domain/policy/WaitListTokenPolicy.java
@@ -5,10 +5,10 @@ import org.springframework.stereotype.Component;
 @Component
 public class WaitListTokenPolicy {
 
-    private final int DEFAULT_SESSION_MINUTES = 5;
+    private final long DEFAULT_SESSION_MINUTES = 5;
     private final int DEFAULT_MAX_CAPACITY = 10;
 
-    public int getSessionMinutes() {
+    public long getSessionMinutes() {
         return DEFAULT_SESSION_MINUTES;
     }
